apps/workers/pyproject.toml
[project]
name = "stepsquad-workers"
version = "0.4.0"
requires-python = ">=3.11"
dependencies = [
  "pydantic>=2.8.0",
  "google-cloud-pubsub>=2.21.0",
  "google-cloud-firestore>=2.16.0",
  "google-cloud-bigquery>=3.25.0"
]
